+++ src/tooling/piral-cli/src/apps/index.ts
@@ -9,6 +9,7 @@ export * from './new-pilet';
 export * from './upgrade-pilet';
 export * from './add-piral-instance-pilet';
 export * from './remove-piral-instance-pilet';
+export * from './run-emulator-piral';
 export * from './debug-pilet';
 export * from './build-pilet';
 export * from './pack-pilet';